+++ ms-personas/src/main/java/com/banco/digital/ms_personas/service/AddressService.java
@@ -4,6 +4,6 @@ import com.banco.digital.ms_personas.model.Address;
 import com.banco.digital.ms_personas.model.User;
 import com.banco.digital.ms_personas.request.UserRegisterRequest;
 
-public interface AddressService extends BaseService<Address, Long>{
+public interface AddressService extends BaseService<Address, Integer>{
     Address generateAddress(UserRegisterRequest userRegisterRequest, User user);
 }
