+++ ms-personas/src/main/java/com/banco/digital/ms_personas/service/UserTypeService.java
@@ -2,6 +2,6 @@ package com.banco.digital.ms_personas.service;
 
 import com.banco.digital.ms_personas.model.UserType;
 
-public interface UserTypeService extends BaseService<UserType, Long> {
+public interface UserTypeService extends BaseService<UserType, Integer> {
     UserType findByDescription(String name);
 }
